+++ crates/rhdl/src/lib.rs
@@ -3,4 +3,5 @@ pub use rhdl_bits as bits;
 pub use rhdl_core as core;
 pub use rhdl_core::const_max;
 pub use rhdl_trace_type as rtt;
+pub use rhdl_vlog as vlog;
 pub use serde_json;