+++ doc/book/src/code/src/count_ones.rs
@@ -213,6 +213,7 @@ set_io leds[3] E14
     // ANCHOR: ones-step-12
     #[test]
     fn test_base_timing() -> miette::Result<()> {
+        skip_without_icestorm!();
         let uut = OneCounter {};
         let timing = rhdl_toolchains::icestorm::IceStorm::new(
             "hx8k",
@@ -342,6 +343,7 @@ pub mod step_6 {
     // ANCHOR: ones-step-18
     #[test]
     fn test_timing_divided() -> miette::Result<()> {
+        skip_without_icestorm!();
         let uut = OneCounterDivided {};
         let timing = rhdl_toolchains::icestorm::IceStorm::new(
             "hx8k",
@@ -410,6 +412,7 @@ pub mod step_7 {
     // ANCHOR: ones-step-20
     #[test]
     fn test_timing_divided_four() -> miette::Result<()> {
+        skip_without_icestorm!();
         let uut = OneCounterDividedFour {};
         let timing = rhdl_toolchains::icestorm::IceStorm::new(
             "hx8k",